% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAGEToken.R
\name{getAGEToken}
\alias{getAGEToken}
\title{Get a security token to use to VIEW data in Spatial Hub - ArcGIS Enterprise - Portal}
\usage{
getAGEToken()
}
\value{
The token string to be used in getWebServiceData
}
\description{
Get a security token to access to restricted content in Cefas ArcGIS Enterprise .The token has to be used within the getWebServiceData function
The token can be used to access to secure resources that are needed for R and Python analyst to access to data
}
\examples{

token_string = getAGEToken ()

getWebServiceData ( web_service=  "https://giserver.cefas.co.uk/devserver/rest/services/fishing_activity/Landing_composition_statistics_by_species_ICES_rectangles_and_ports_2009_to_2021/MapServer",
                   layer_index = 8,
                   where_clause = "ft_lyear = 2021 AND le_spe = 'COD'",
                   output_fields = "*",
                   spatial_layer = TRUE ,
                   token_str = token_string )


}
