% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCsquareCode.R
\name{getCsquareCode}
\alias{getCsquareCode}
\title{Get the CSquare geocode giving a pair of coordinates longitude and latitude of a given location}
\usage{
getCsquareCode(lon, lat, degrees)
}
\arguments{
\item{lon}{latitude of the location}

\item{lat}{longitude of the location}

\item{degrees}{size of the C-Square cell in decimal degrees . Default value is 0.05}
}
\value{
A CSquare geocode list for the inputed pair of coordinates
}
\description{
Get the Csquare geocode for a given pair of coordinates longitue and latide of a given geographycal location
The Csquare geocoding allows to have a unique identifier for each of the locations. The CSquare code allows to join the dataframe to a spatial grid using CSquare geocode and same spatial resolution grid cell
The pair of coordinates must be given in  a unprojected coordinate system ( WGS1984 - EPSG: 4326 )
this function is taken from VMSTools package
}
\details{
To create a spatial grid with CSquare geocode labels use the function getCsquareGrid
}
