% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCsquareGrid.R
\name{getCsquareGrid}
\alias{getCsquareGrid}
\title{Create a grid with CSquare notation to aggregate spatial data in Spatial Hub metadata}
\usage{
getCsquareGrid(spatial_layer, cell_size = 0.05, projection = 4326)
}
\arguments{
\item{spatial_layer}{Spatial dataset in SF format that is intended to be gridded using C-Squares}

\item{cell_size}{longitude of the location}

\item{projection}{Projection required for the calculation of C-Square grid cells areas. Default 4326}

\item{degrees}{size of the C-Square cell in decimal degrees . Default value is 0.05}
}
\value{
A spatial grid for a required extension and with a Csquare label for each grid cell
}
\description{
Create a spatial grid with a given cell resolution and with a related Csquare geocode for each of the grid cells
The Csquare geocoding allows to have a unique identifier for each of the locations. The grid is created in a unprojected coordinate system ( WGS1984 - EPSG: 4326 )
}
\examples{

token_string = getCsquareGrid (spatial_layer  , cell_size =0.05, projection = 4326)



}
