% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWebServiceData.R
\name{getWebServiceData}
\alias{getWebServiceData}
\title{Get the Spatial Hub web services data into R objetcs}
\usage{
getWebServiceData(
  web_service,
  layer_index,
  where_clause = "1=1",
  output_fields = "*",
  spatial_layer = TRUE,
  token_str = NULL
)
}
\arguments{
\item{web_service}{The URL of the Spatial Hub - ArcGIS Server web service .}

\item{layer_index}{ArcGIS Server API's could include several layers. The index number of the required layer is here provided}

\item{where_clause}{A SQL format WHERE clause. The SQL condition filters the data to be extracted. To extract all the data use "1=1" string
\itemize{
  \item Extract all data  - where_clause = "1=1"
  \item Filter data using SQL query - where_clause = "le_spe = 'COD' AND ft_lyear = 2021"
}}

\item{output_fields}{An array with the list of the fields that are reuiqred with the output}

\item{spatial_layer}{A TRUE / FALSE parameter to return the output data in  R SF spatial format  or as a R Data Frame format}

\item{token}{if the api is protected a token must be provided}
}
\value{
The data extracted as a Data Frame or Spatial SF format
}
\description{
Get Spatial Hub web services API's into R objects. The function allows to access the data as a R Data Frame or spatial SF object
The Spatial Hub - ArcGIS Portal API's allows  the use of SQL query to extract only the required data
}
\examples{


getWebServiceData ( web_service=  "https://giserver.cefas.co.uk/devserver/rest/services/fishing_activity/Landing_composition_statistics_by_species_ICES_rectangles_and_ports_2009_to_2021/MapServer",
                   layer_index = 8,
                   where_clause = "ft_lyear = 2021 AND le_spe = 'COD'",
                   output_fields = "*",
                   spatial_layer = TRUE )


}
